refactor: move response type from model/init.go to reponse/reponse.go
