+++ backend/api/participant.go
@@ -15,9 +15,9 @@ import (
 // @Accept       json
 // @Produce      json
 // @Param   	 competitionID 	 formData int true "competition id"
-// @Success      200  {object}  model.Response "success"
-// @Failure      400  {object}  model.Response "cannot parse competitionID | participant exists"
-// @Failure      400  {object}  model.Response "no user/competition found"
+// @Success      200  {object}  response.Response "success"
+// @Failure      400  {object}  response.Response "cannot parse competitionID | participant exists"
+// @Failure      400  {object}  response.Response "no user/competition found"
 // @Router       /participant/ [post]
 func JoinInCompetition(c *gin.Context) {
 	userID := pkg.QuerySession(c, "id").(uint)
